requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.7.0
pandas>=2.0.0
openpyxl>=3.1.0
flask>=3.0.0
gunicorn>=21.2.0


